+++ scrypto/src/resource/resource_def.rs
@@ -41,7 +41,7 @@ impl ResourceDef {
         metadata: HashMap<String, String>,
         flags: u64,
         mutable_flags: u64,
-        authorities: HashMap<Address, u16>,
+        authorities: HashMap<Address, u64>,
         initial_supply: Option<NewSupply>,
     ) -> (ResourceDef, Option<Bucket>) {
         let input = CreateResourceInput {
